Revert "Fix minRange and maxRange when using stepsArray (issue #60)"
I decided that this is not the right approach to take.

It is better for the minRange and maxRange options to be expressed
in terms of indexes into stepsArray, as they have been so far.

This is consistent with the convention used by stepsArray so far,
which is to keep all internal state of slider as if it bound to an index
into stepsArray - a range of 0 to stepsArray.length()-1.

The values of stepsArray only come into play when binding value and
highValue inputs/outputs of the slider, and are otherwise not taken into
account when controlling the slider.

This reverts commit 033d59bd16e68aed3a39c2797db34c53b61378c6.